requirements.txt
fastapi
uvicorn[standard]
pyserial
pyserial-asyncio
